Move helper functions from main AWS stack into a separate utils file
